Cleaned the code a bit (#60)
* :art: Cleaned the code a bit

Removed try except block for checking if the config exists, replaced by `os.path.isfile`. Also added checks for when to write to the config file. This is because previously we would keep writing the same token to the same config each run. Now it only writes the token to the config if the config doesn't exist yet. Also did some minor things like change around comments.

* :rewind: Revert some changes